fix ducks format
