--- src/redux/modules/rootReducer.js
@@ -1,9 +1,0 @@
-import { combineReducers } from 'redux';
-
-import films from './films';
-import seats from './seats';
-
-export default combineReducers({
-  films,
-  seats,
-});
